Refactor REST API.
