--- lift-web/public/javascripts/forms.js
@@ -1,7 +1,0 @@
-function formSubmit() {
-    document.getElementById("setForm").addEventListener("submit", function (event) {
-        event.preventDefault()
-    });
-}
-
-module.exports = { formSubmit };
