+++ lift-web/views/partials/footer.ejs
@@ -1,4 +1,4 @@
-<footer class="footer bg-dark py-3 mt-auto">
+<footer class="footer py-3 mt-auto">
     <div class=" container">
         <span class="text-light">&copy; Lift 2022</span>
     </div>
